Cleanup, Optimization, Localization, Config, License
Added .editorconfig, to setup a coding standart for the repo.

__resource.lua Changede the loading order, mysql-async was loaded after server script, changed to before, added locales, and config to server and client side.

Created folders for client and server script, renamed both to main.lua.

client\main.lua Cleaned up the code, moved variables to the top, changed 3600 to Config.GsrTimer to allow easy change of "GSR Memory", added chat help text, removed all notification events, added one single event for notifications, with vars text and type.

server\main.lua Cleaned up the code, changed the way /gsr is handled, now using es:addCommand (To avoid chat saying command does not exist), added error handling if ID is missing, or player is not online. Changed onMySQLReady to MySQL.ready (required mysql-async 3.x.x), added localization.

Localization added, so translations can be done with ease.

Config file added, for changing settings, insted of in script config options.

locales\en.lua English base translation file.

gsr.sql cleaned up, INT changed from INT(250) - Invalid to INT(11) - Valid.

Removed readme.txt added README.md with markdown.

Added a valid license option, current "license" is not valid and can be avoided if a person wants, so GNU v3 is a better option.

Code testet and is working.
