+++ .editorconfig
@@ -1,0 +1,9 @@
+root = true
+
+[*]
+indent_size = 4
+indent_style = tab
+end_of_line = crlf
+charset = utf-8
+trim_trailing_whitespace = true
+insert_final_newline = true
