+++ __resource.lua
@@ -1,10 +1,16 @@
 resource_manifest_version '44febabe-d386-4d18-afbe-5e627f4af937'
 
-client_scripts {
-	"client.lua"
+server_scripts {
+	'@mysql-async/lib/MySQL.lua',
+	'@es_extended/locale.lua',
+	'locales/en.lua',
+	'config.lua',
+	'server/main.lua'
 }
 
-server_scripts {
-	"server.lua",
-	"@mysql-async/lib/MySQL.lua"
+client_scripts {
+	'@es_extended/locale.lua',
+	'locales/en.lua',
+	'config.lua',
+	"client/main.lua"
 }
