+++ config.lua
@@ -1,0 +1,5 @@
+Config                          = {}
+Config.Locale                   = 'en'
+
+Config.waterClean               = true -- Set to false if you dont want water to clean off GSR from people who shot
+Config.GsrTime                  = 60 * 60 -- Change The first number if you want the GSR to be auto removed faster output is minutes.
