+++ gsr.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS `gsr` (
-  `identifier` varchar(250) COLLATE utf8_unicode_ci NOT NULL,
-  `time` int(250) NOT NULL,
+  `identifier` varchar(250) NOT NULL,
+  `time` int(11) NOT NULL,
   PRIMARY KEY (`identifier`)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
+);
