+++ locales/en.lua
@@ -1,0 +1,9 @@
+Locales['en'] = {
+  -- Chat help text
+  ['help_gsr'] = 'Test person for GSR',
+  ['help_gsr_value'] = 'ID',
+  -- GSR Messages
+  ['gsr_cleaned'] = 'You hopped in the water and cleaned off the Gunshot Residue.',
+  ['gsr_positive'] = 'Traces of gunpowder were found.',
+  ['gsr_negative'] = 'No traces of gunpowder were found',
+}
